Flatten comorbidities subfolders into single .py files per comorbidity
Replace the deeply nested tbsim/comorbidities/hiv/ and
tbsim/comorbidities/malnutrition/ directories with single consolidated
modules at tbsim/comorbidities/hiv.py and tbsim/comorbidities/malnutrition.py.
Updates __init__.py imports and docs accordingly. Closes #294.
