+++ tbsim/__init__.py
@@ -3,11 +3,8 @@ DATADIR = os.path.join(os.path.dirname(__file__), 'data')
 
 from .tb_lshtm import *
 from .tb import *
-from .comorbidities.malnutrition.malnutrition import *
-from .comorbidities.hiv.hiv import *
-from .comorbidities.hiv.intervention import *
-from .comorbidities.hiv.tb_hiv_cnn import *
-from .comorbidities.malnutrition.tb_malnut_cnn import *
+from .comorbidities.hiv import *
+from .comorbidities.malnutrition import *
 from .interventions import *
 from .networks import *
 from .analyzers import *
